Comment out unclear meta refresh test.
The following HTML element previously caused a reload of "/login":

<meta http-equiv="refresh" content="0;'URL=/login'">

Currently, no reload happens, probably as a result of a change in
mechanize's behavior. However, it's not clear what the correct action
would be, if any. For example, Firefox attempts to reload
"/URL=/login", a non-existent page. Acceptable format for the content
attribute seems ill-specified at best.